+++ src/galleria/cli.py
@@ -5,11 +5,14 @@ Author: Marcus Grant
 Created: 2026-08-24
 License: AGPL-3.0-or-later
 """
+
 import click
 
 from galleria.command.build import build
+from galleria.command.derive import derive
 from galleria.command.process_photos import process_photos
 from galleria.command.serve import serve
+from galleria.command.validate import validate
 
 
 @click.group()
@@ -19,5 +22,8 @@ def cli():
 
 
 cli.add_command(process_photos, name="process-photos")
+cli.add_command(validate, name="validate")
+cli.add_command(derive, name="derive")
 cli.add_command(build, name="build")
 cli.add_command(serve, name="serve")
+